// shared/addresses/types.ts
export interface AddressWordBase {
  nodeType: "word";
  value: string;
}

export interface AddressNodeWithUnclassifiedWord extends AddressWordBase {
  wordType: "unclassified";
}
// ...
export interface WordReplacementConfig {
  detached?: boolean;
  from: string | string[];
  to: string;
  silenceNormalizationWarning?: boolean;
}

export interface WordReplacementDirective {
  detached: boolean;
  from: AddressNodeWithUnclassifiedWord[];
  to: AddressNodeWithUnclassifiedWord[];
}

export interface WordReplacementDirectiveTree {
  directive?: WordReplacementDirective;
  subtreeByWordValue: Record<string, WordReplacementDirectiveTree>;
}
// ...
export interface AddressNormalizationConfig
  extends AddressCleaningConfig,
    AddressStandardizationConfig {}

export interface AddressBeautificationConfig
  extends AddressNormalizationConfig {
  canonicalSpellings?: string[];
}

export interface RawAddressHandlingConfig {
  canonicalSpellings?: string[];
  defaultRegion?: string;
  wordReplacements?: WordReplacementConfig[];
}

export type AddressHandlingConfig = AddressBeautificationConfig;